+++ analysis.py
@@ -10,16 +10,32 @@ import matplotlib.pyplot as plt
 from exceptions import error_popup, warning_popup
 
 
-def test():
-    print('test')
-
 def marker_euclidean_distance(p1x, p1y, p2x, p2y):
+    """calculates the euclidean distance between 2 points in the 2D space
+
+    Args:
+        p1x (int): x coord of point 1
+        p1y (int): y coord of point 1
+        p2x (int): x coord of point 2
+        p2y (int): y coord of point 2
+
+    Returns:
+        float: the euclidean distance between the two given points
+    """    
     return np.sqrt((p2x - p1x)**2 + (p2y - p1y)**2)
 
 def plot_data(x, y, plot_args):
+    """util function to handle plotting and formatting of the plots
+
+    Args:
+        x (pd.Dataframe/Series, np.Array, list, etc): _description_
+        y (pd.Dataframe/Series, np.Array, list, etc): _description_
+        plot_args (dict): dictionary of plot customization arguments
+
+    Returns:
+        (plt.figure, plt.axes): the figure and axes objects created and modified in this function
+    """    
     fig, ax = plt.subplots()
-
-
 
     # plot data, adding legend depending on plot args
     if plot_args['has_legend']:
@@ -36,6 +52,11 @@ def plot_data(x, y, plot_args):
     return fig, ax
 
 def analyze_marker_deltas():
+    """plots euclidean distance between tracked fiducial marker data
+    reads data from 'output/Tracking_Output.csv' which is created in the marker tracking process
+    saves plot to the 'figures' folder
+    """    
+    print("Analyzing tracked marker distances...")
     df = pd.read_csv("output/Tracking_Output.csv") # open csv created/modified from marker tracking process
     print(df.head())
 
@@ -49,7 +70,6 @@ def analyze_marker_deltas():
     m1_df = df[df['Tracker'] == 1]
     m2_df = df[df['Tracker'] == 2]
     time = df['Time(s)'].unique()
-    print(time)
     m1_x = m1_df['x (px)'].values
     m1_y = m1_df['y (px)'].values
     m2_x = m2_df['x (px)'].values
@@ -73,7 +93,33 @@ def analyze_marker_deltas():
 
     fig, ax = plot_data(time, marker_deltas, plot_args)
     fig.savefig("figures/marker_deltas.png")
-    plt.show()
+    print("Done")
+
+
+def analyze_necking_point():
+    """plots necking point data, x location of necking point against time
+    reads from 'output/Necking_Point_Output.csv' which is created in the necking point tracking process
+    saves plot in 'figures' folder
+    """    
+    print("Analyzing necking point...")
+    df = pd.read_csv("output/Necking_Point_Output.csv") # open csv created/modified from marker tracking process
+    print(df.head())
+
+    time = df['Time(s)'].values
+    necking_pt_x = df['x at necking point (px)'].values
+
+    plot_args = {
+        'title': 'Necking Point Detection',
+        'x_label': 'Time (s)',
+        'y_label': 'Horizontal coordinate of necking point (px)',
+        'data_label': '',
+        'has_legend': False
+    }
+
+    fig, ax = plot_data(time, necking_pt_x, plot_args)
+    fig.savefig("figures/necking_point_location.png")
+
+    print("Done")
 
 if __name__=='__main__':
     analyze_marker_deltas()